Update quantity input type for procurement reporting
Changed the quantity field from float to integer in the `procure_from_requirements` function within `routes/report_routes.py` to ensure whole number quantities are processed for procurement.

+++ routes/report_routes.py
@@ -163,7 +163,7 @@ def procure_from_requirements():
             try:
                 product_id = int(request.form.get(f'items[{i}][product_id]'))
                 supplier_id = int(request.form.get(f'items[{i}][supplier_id]'))
-                quantity = float(request.form.get(f'items[{i}][quantity]'))
+                quantity = int(request.form.get(f'items[{i}][quantity]'))
                 unit_price = float(request.form.get(f'items[{i}][unit_price]'))
                 
                 existing_po_count = conn.execute(
